move util stuff into util
