+++ vsc-extension/src/extension.ts
@@ -1,24 +1,15 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
-import { relative } from "path";
 import * as vscode from "vscode";
 import {
   MaybeSelectedFunction,
   SourceFileMap,
 } from "../src-shared/source-info";
-import { focusNode, getEditorSelectedFunction } from "./editor-utils";
 import { ImaginaryFunctionProvider } from "./function-tree-provider";
 import { ImaginaryMessageRouter } from "./imaginary-message-router";
-import { registerWebView } from "./react-webview-provider";
-import { removeFile, updateFile } from "./source-utils";
-
-export function getRelativePathToProject(absPath: string) {
-  const projectPath = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
-  if (projectPath) {
-    return relative(projectPath, absPath);
-  }
-  return absPath;
-}
+import { focusNode, getEditorSelectedFunction } from "./util/editor";
+import { registerWebView } from "./util/react-webview-provider";
+import { removeFile, updateFile } from "./util/source";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
