+++ vsc-extension/src/util/ast.ts
