+++ vsc-extension/src/util/source.ts
@@ -1,8 +1,8 @@
 import * as ts from "typescript";
 import * as vscode from "vscode";
-import { SourceFileMap } from "../src-shared/source-info";
-import { findFunctions } from "./ast-utils";
-import { getRelativePathToProject } from "./extension";
+import { SourceFileMap } from "../../src-shared/source-info";
+import { findFunctions } from "./ast";
+import { getRelativePathToProject } from "./editor";
 
 export function removeFile(
   document: vscode.TextDocument,
